Move segmentation functions into separate service
