+++ src/main/kotlin/simplecolocalization/SimpleColocalization.kt
@@ -4,23 +4,14 @@ import ij.IJ
 import ij.ImagePlus
 import ij.WindowManager
 import ij.gui.MessageDialog
-import ij.gui.Roi
-import ij.measure.Measurements
-import ij.measure.ResultsTable
-import ij.plugin.ChannelSplitter
 import ij.plugin.ZProjector
-import ij.plugin.filter.BackgroundSubtracter
-import ij.plugin.filter.EDM
-import ij.plugin.filter.MaximumFinder
-import ij.plugin.filter.ParticleAnalyzer
-import ij.plugin.filter.RankFilters
 import ij.plugin.frame.RoiManager
-import ij.process.ImageConverter
 import java.io.File
 import net.imagej.Dataset
 import net.imagej.ImageJ
 import org.scijava.ItemVisibility
 import org.scijava.command.Command
+import org.scijava.log.LogService
 import org.scijava.plugin.Parameter
 import org.scijava.plugin.Plugin
 import org.scijava.table.DefaultGenericTable
@@ -40,6 +31,12 @@ import org.scijava.widget.NumberWidget
  */
 @Plugin(type = Command::class, menuPath = "Plugins > Simple Colocalization")
 class SimpleColocalization : Command {
+
+    @Parameter
+    private lateinit var logService: LogService
+
+    @Parameter
+    private lateinit var cellSegmentationService: CellSegmentationService
 
     /**
      * Entry point for UI operations, automatically handling both graphical and
@@ -84,7 +81,7 @@ class SimpleColocalization : Command {
      * Used during the cell identification stage to reduce overlapping cells
      * being grouped into a single cell.
      *
-     * TODO (#5): Figure out what this value should be.
+     * TODO(#5): Figure out what this value should be.
      */
     @Parameter(
         label = "Largest Cell Diameter",
@@ -118,49 +115,6 @@ class SimpleColocalization : Command {
         uiService.show(table)
     }
 
-    /**
-     * Perform pre-processing on the image to remove background and set cells to white.
-     */
-    private fun preprocessImage(image: ImagePlus) {
-        // Convert to grayscale 8-bit.
-        ImageConverter(image).convertToGray8()
-
-        // Remove background.
-        BackgroundSubtracter().rollingBallBackground(
-            image.channelProcessor,
-            largestCellDiameter,
-            false,
-            false,
-            false,
-            false,
-            false
-        )
-
-        // Threshold grayscale image, leaving black and white image.
-        image.channelProcessor.autoThreshold()
-
-        // Despeckle the image using a median filter with radius 1.0, as defined in ImageJ docs.
-        // https://imagej.nih.gov/ij/developer/api/ij/plugin/filter/RankFilters.html
-        RankFilters().rank(image.channelProcessor, 1.0, RankFilters.MEDIAN)
-
-        // Apply Gaussian Blur to group larger speckles.
-        image.channelProcessor.blurGaussian(gaussianBlurSigma)
-
-        // Threshold image to remove blur.
-        image.channelProcessor.autoThreshold()
-    }
-
-    /**
-     * Segment the image into individual cells, overlaying outlines for cells in the image.
-     *
-     * Uses ImageJ's Euclidean Distance Map plugin for performing the watershed algorithm.
-     * Used as a simple starting point that'd allow for cell counting.
-     */
-    private fun segmentImage(image: ImagePlus) {
-        // TODO (#7): Review and improve upon simple watershed.
-        EDM().toWatershed(image.channelProcessor)
-    }
-
     /** Runs after the parameters above are populated. */
     override fun run() {
         var image = WindowManager.getCurrentImage()
@@ -181,14 +135,14 @@ class SimpleColocalization : Command {
     private fun process(image: ImagePlus) {
         val originalImage = image.duplicate()
 
-        preprocessImage(image)
-        segmentImage(image)
+        cellSegmentationService.preprocessImage(image, largestCellDiameter, gaussianBlurSigma)
+        cellSegmentationService.segmentImage(image)
 
         val roiManager = RoiManager.getRoiManager()
-        val cells = identifyCells(roiManager, image)
-        markCells(originalImage, cells)
+        val cells = cellSegmentationService.identifyCells(roiManager, image)
+        cellSegmentationService.markCells(originalImage, cells)
 
-        val analysis = analyseCells(originalImage, cells)
+        val analysis = cellSegmentationService.analyseCells(originalImage, cells)
 
         showCount(analysis)
         showPerCellAnalysis(analysis)
@@ -196,77 +150,8 @@ class SimpleColocalization : Command {
         originalImage.show()
     }
 
-    /**
-     * Select each cell identified in the segmented image in the original image.
-     *
-     * We use [ParticleAnalyzer] instead of [MaximumFinder] as the former highlights the shape of the cell instead
-     * of just marking its centre.
-     */
-    private fun identifyCells(roiManager: RoiManager, segmentedImage: ImagePlus): Array<Roi> {
-        ParticleAnalyzer.setRoiManager(roiManager)
-        ParticleAnalyzer(
-            ParticleAnalyzer.SHOW_NONE or ParticleAnalyzer.ADD_TO_MANAGER,
-            Measurements.ALL_STATS,
-            ResultsTable(),
-            0.0,
-            Double.MAX_VALUE
-        ).analyze(segmentedImage)
-        return roiManager.roisAsArray
-    }
-
-    /**
-     * Mark the cell locations in the image.
-     */
-    private fun markCells(image: ImagePlus, rois: Array<Roi>) {
-        for (roi in rois) {
-            roi.image = image
-        }
-    }
-
     data class CellAnalysis(val area: Int, val channels: List<ChannelAnalysis>)
     data class ChannelAnalysis(val name: String, val mean: Int, val min: Int, val max: Int)
-
-    /**
-     * Analyses the channel intensity of the cells.
-     */
-    private fun analyseCells(image: ImagePlus, highlightedCells: Array<Roi>): Array<CellAnalysis> {
-        // Split the image into multiple grayscale images (one for each channel).
-        val channelImages = ChannelSplitter.split(image)
-        val numberOfChannels = channelImages.size
-
-        val analyses = arrayListOf<CellAnalysis>()
-        for (cell in highlightedCells) {
-            var area = 0
-            val sums = MutableList(numberOfChannels) { 0 }
-            val mins = MutableList(numberOfChannels) { Integer.MAX_VALUE }
-            val maxs = MutableList(numberOfChannels) { Integer.MIN_VALUE }
-            val containedCells = cell.containedPoints
-            containedCells.forEach { point ->
-                area++
-                for (channel in 0 until numberOfChannels) {
-                    // pixelData is of the form [value, 0, 0, 0] because ImageJ.
-                    val pixelData = channelImages[channel].getPixel(point.x, point.y)
-                    sums[channel] += pixelData[0]
-                    mins[channel] = Integer.min(mins[channel], pixelData[0])
-                    maxs[channel] = Integer.max(maxs[channel], pixelData[0])
-                }
-            }
-            val channels = mutableListOf<ChannelAnalysis>()
-            for (channel in 0 until numberOfChannels) {
-                channels.add(
-                    ChannelAnalysis(
-                        channelImages[channel].title,
-                        sums[channel] / area,
-                        mins[channel],
-                        maxs[channel]
-                    )
-                )
-            }
-            analyses.add(CellAnalysis(area, channels))
-        }
-
-        return analyses.toTypedArray()
-    }
 
     /**
      * Displays the resulting cell analysis as a results table.
